feat: run redundant_test_main_exit rule when go version is greater than or equal to 1.15
