+++ testdata/go1.15/go.mod
@@ -1,0 +1,3 @@
+module github.com/mgechev/revive/testdata
+
+go 1.15
