+++ testdata/go1.15/redundant_test_main_exit.go
