+++ testdata/redundant_test_main_exit_test.go
@@ -3,7 +3,6 @@ package fixtures
 import (
 	"fmt"
 	"os"
-	"syscall"
 	"testing"
 )
 
@@ -11,8 +10,8 @@ func TestMain(m *testing.M) {
 	setup()
 	i := m.Run()
 	teardown()
-	os.Exit(i)      // MATCH /redundant call to os.Exit in TestMain function, the test runner will handle it automatically as of Go 1.15/
-	syscall.Exit(i) // MATCH /redundant call to syscall.Exit in TestMain function, the test runner will handle it automatically as of Go 1.15/
+	// must not match because the go version of this module is less than 1.15
+	os.Exit(i)
 }
 
 func setup() {
@@ -25,14 +24,4 @@ func teardown() {
 
 func Test_function(t *testing.T) {
 	t.Error("Fail")
-}
-
-func Test_os_exit(t *testing.T) {
-	// must not match because this is not TestMain function
-	os.Exit(1)
-}
-
-func Test_syscall_exit(t *testing.T) {
-	// must not match because this is not TestMain function
-	syscall.Exit(1)
 }
